requirements.txt
torch
transformers
pandas
numpy
pyyaml
keyboard
pytest
# optional on Windows notifications:
# win10toast
